+++ src/main/java/com/forgestove/create_cyber_goggles/content/config/CCGConfig.java
@@ -1,19 +1,13 @@
 package com.forgestove.create_cyber_goggles.content.config;
-import com.forgestove.create_cyber_goggles.content.util.*;
 import com.terraformersmc.modmenu.api.*;
 import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.serializer.Toml4jConfigSerializer;
 import net.fabricmc.api.*;
 @Environment(EnvType.CLIENT)
 public class CCGConfig implements ModMenuApi {
-	private static CCGConfigData config;
-	public static CCGConfigData get() {
-		if (config == null) SafeRun.run(() -> config =
-			AutoConfig.register(CCGConfigData.class, Toml4jConfigSerializer::new).getConfig());
-		return config;
-	}
+	public static CCGConfigData config;
 	public static void register() {
-		SafeRun.run(() -> config = AutoConfig.register(CCGConfigData.class, Toml4jConfigSerializer::new).getConfig());
+		config = AutoConfig.register(CCGConfigData.class, Toml4jConfigSerializer::new).getConfig();
 	}
 	@Override
 	public ConfigScreenFactory<?> getModConfigScreenFactory() {