+++ src/main/java/com/forgestove/create_cyber_goggles/mixin/armor/DivingBootsItemMixin.java
@@ -1,5 +1,5 @@
 package com.forgestove.create_cyber_goggles.mixin.armor;
-import com.forgestove.create_cyber_goggles.content.config.CCGConfig;
+import com.forgestove.create_cyber_goggles.content.config.*;
 import com.simibubi.create.content.equipment.armor.DivingBootsItem;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Mixin;
@@ -9,6 +9,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public abstract class DivingBootsItemMixin {
 	@Inject(method = "isWornBy", at = @At("HEAD"), remap = false, cancellable = true)
 	private static void isWornBy(@NotNull CallbackInfoReturnable<Boolean> returnable) {
-		if (CCGConfig.get().armor.removeDivingBootsAffect) returnable.setReturnValue(false);
+		if (CCGConfig.config.armor.removeDivingBootsAffect) returnable.setReturnValue(false);
 	}
 }