+++ src/main/java/com/forgestove/create_cyber_goggles/mixin/goggles/GogglesItemMixin.java
@@ -1,5 +1,5 @@
 package com.forgestove.create_cyber_goggles.mixin.goggles;
-import com.forgestove.create_cyber_goggles.content.config.CCGConfig;
+import com.forgestove.create_cyber_goggles.content.config.*;
 import com.simibubi.create.content.equipment.goggles.GogglesItem;
 import net.minecraft.client.Minecraft;
 import org.spongepowered.asm.mixin.Mixin;
@@ -14,10 +14,9 @@ public abstract class GogglesItemMixin {
 			returnable.setReturnValue(false);
 			return;
 		}
-		var gameMode = mc.gameMode;
-		if (gameMode == null) return;
-		var goggles = CCGConfig.get().goggles;
-		if (!switch (gameMode.getPlayerMode()) {
+		if (mc.gameMode == null) return;
+		var goggles = CCGConfig.config.goggles;
+		if (!switch (mc.gameMode.getPlayerMode()) {
 			case SURVIVAL -> goggles.enableInSurvival;
 			case CREATIVE -> goggles.enableInCreative;
 			case SPECTATOR -> goggles.enableInSpectator;