+++ src/main/java/com/forgestove/create_cyber_goggles/mixin/other/CreateJEIMixin.java
@@ -1,4 +1,4 @@
-package com.forgestove.create_cyber_goggles.mixin.jei;
+package com.forgestove.create_cyber_goggles.mixin.other;
 import com.forgestove.create_cyber_goggles.content.config.CCGConfig;
 import com.simibubi.create.compat.jei.CreateJEI;
 import org.spongepowered.asm.mixin.Mixin;
@@ -13,6 +13,6 @@ public abstract class CreateJEIMixin {
 	), remap = false, index = 1
 	)
 	private int loadCategories(int height) {
-		return CCGConfig.get().other.nonrandomScrap ? height + 40 : height;
+		return CCGConfig.config.other.nonrandomScrap ? height + 40 : height;
 	}
 }