+++ src/main/java/com/forgestove/create_cyber_goggles/mixin/other/NbtAccounterMixin.java
@@ -1,4 +1,4 @@
-package com.forgestove.create_cyber_goggles.mixin.nbt;
+package com.forgestove.create_cyber_goggles.mixin.other;
 import com.forgestove.create_cyber_goggles.content.config.CCGConfig;
 import com.forgestove.create_cyber_goggles.content.util.SafeRun;
 import net.minecraft.nbt.NbtAccounter;
@@ -9,6 +9,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public abstract class NbtAccounterMixin {
 	@Inject(method = "accountBytes(J)V", at = @At("HEAD"), cancellable = true)
 	public void accountBytes(CallbackInfo callbackInfo) {
-		SafeRun.run(() -> {if (CCGConfig.get().other.nbtFix) callbackInfo.cancel();});
+		SafeRun.run(() -> {if (CCGConfig.config.other.nbtFix) callbackInfo.cancel();});
 	}
 }