+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Reworked configurations and translation.
+- Reworked mixin